# mvc-test0-openapi.yaml
openapi: 3.0.1
info:
  title: mvc-test0 openapi
  description: test swagger annotations on endpoints
  license:
    name: personal ip license
    url: https://ip.me
  version: 1.5.12
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: api11
  description: "the API for Api1: tag1"
- name: api12
  description: "the API for Api1: tag2"
paths:
  /test_api1/add:
    post:
      tags:
      - controller-1
      summary: "api1: add data"
      description: Add new data
      operationId: addData
      requestBody:
        description: Created data object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api1_data'
          application/xml:
            schema:
              $ref: '#/components/schemas/api1_data'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/api1_data'
        required: true
      responses:
        "405":
          description: Invalid input
        "200":
          description: Successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/api1_data'
            application/json:
              schema:
                $ref: '#/components/schemas/api1_data'
  /test_api1/m12/{param1}:
    get:
      tags:
      - api12
      summary: api1:get mapping2
      description: m12 mapping with pathvar description
      operationId: m12
      parameters:
      - name: param1
        in: path
        description: some path variable
        required: true
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        "400":
          description: Invalid status value
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
  /test_api1/all:
    get:
      tags:
      - api11
      summary: api1:get all datas
      description: return all data
      operationId: all
      responses:
        "200":
          description: successful operation
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        "400":
          description: Invalid status value
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
components:
  schemas:
    api1_data:
      type: object
      properties:
        first_property:
          type: string
          description: api1_data prop1
        second_property:
          type: string
          description: api1_data prop2
      description: data object for api1